feat: enhance revenue dashboard with service metrics and improve reviews section
- Updated Revenue component to include top services sold and their revenue.
- Modified RevenueData type to accommodate new service-related fields.
- Adjusted summary statistics to reflect service revenue and products sold.
- Enhanced Reviews component to highlight pending reviews and improve UI.
- Added total reviews count to OwnerPitchSummaryDto.
- Introduced TopServiceDto for better service tracking in OwnerDashboardDto.
- Implemented CalculateTopServices method to aggregate service data in GetOwnerDashboardQueryHandler.
- Updated mappings and queries to ensure accurate service data retrieval.
- Improved pagination and filtering logic in SearchPitchesQueryHandler.

+++ server/Application/Features/Pitches/MappingProfile.cs
@@ -11,7 +11,7 @@ public class MappingProfile : Profile
         CreateMap<Pitch, PitchDto>()
             .ForMember(dest => dest.TypeDisplay, opt => opt.MapFrom(src => src.Type.ToString()))
             .ForMember(dest => dest.AverageRating, opt => opt.MapFrom(src =>
-                src.Reviews.Any() ? src.Reviews.Average(review => review.Rating) : 0))
+                src.Reviews.Any() ? src.Reviews.Average(review => (decimal)review.Rating) : 0m))
             .ForMember(dest => dest.TotalReviews, opt => opt.MapFrom(src => src.Reviews.Count))
             .ForMember(dest => dest.Address, opt => opt.MapFrom(src => src.SportCenter != null ? src.SportCenter.Address : null))
             .ForMember(dest => dest.Images, opt => opt.MapFrom(src =>
